backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cssselect==1.2.0
dataclasses-json==0.6.7
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
executing==2.1.0
faiss-cpu==1.9.0.post1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.2
jieba3k==0.35.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.145
lxml==5.3.0
lxml_html_clean==0.4.1
marshmallow==3.23.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
orjson==3.10.12
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.21.0
rsa==4.9
sgmllib3k==1.0.0
sidecar==0.7.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.2
stack-data==0.6.3
tenacity==9.0.0
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.0
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.13
yarl==1.18.0
